src/frontend/requirements.txt
httpx==0.28.1
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.49.0
